+++ Vehicle.h
@@ -13,7 +13,7 @@ private:
     GameWorld *m_world;
     SteeringBehaviors *m_steeringBehavior;
 
-    Vector2D<double> m_steeringForce = Vector2D<double> {0.0, 0.0};
+    Vector2D<float> m_steeringForce = Vector2D<float> {0.0, 0.0};
 
     double m_timeElapsed;
 
@@ -21,23 +21,23 @@ public:
     ~Vehicle() override;
 
     Vehicle(GameWorld *m_world,
-            const Vector2D<double> &pos,
-            const Vector2D<double> &scale,
-            const Vector2D<double> &m_velocity,
-            const Vector2D<double> &m_heading,
-            const Vector2D<double> &m_side,
+            const Vector2D<float> &pos,
+            const Vector2D<float> &scale,
+            const Vector2D<float> &m_velocity,
+            const Vector2D<float> &m_heading,
+            const Vector2D<float> &m_side,
             double m_mass,
             double m_maxSpeed,
             double m_maxForce,
             double m_maxTurnRate);
 
-    Vector2D<double> m_wanderTarget = Vector2D<double> {0.0, 0.0};
+    Vector2D<float> m_wanderTarget = Vector2D<float> {0.0, 0.0};
 
-    Vector2D<double> getPos() { return m_pos; }
+    Vector2D<float> getPos() { return m_pos; }
 
-    Vector2D<double> getHeading() { return m_heading; }
+    Vector2D<float> getHeading() { return m_heading; }
 
-    Vector2D<double> getSide() { return m_side; }
+    Vector2D<float> getSide() { return m_side; }
 
     const double getMaxSpeed() { return m_maxSpeed; }
